refactor clustering job
