+++ rrs/clustering/providers.py
@@ -1,0 +1,8 @@
+PROVIDER_MISTRAL = "mistral"
+PROVIDER_ANTHROPIC = "anthropic"
+
+EMBEDDING_BACKEND_ST = "sentence-transformer"
+EMBEDDING_BACKEND_MISTRAL = "mistral"
+
+MODEL_MISTRAL = "mistral-small-2506"
+MODEL_ANTHROPIC = "claude-haiku-4-5-20251001"
